requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
base58==2.1.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coincurve==21.0.0
cytoolz==1.0.1
dnspython==2.7.0
email_validator==2.2.0
eth-hash==0.7.1
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.12
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
parsimonious==0.10.0
pluggy==1.5.0
pycryptodome==3.22.0
pydantic==2.11.2
pydantic_core==2.33.1
pytest==8.3.5
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
toolz==1.0.0
tronpy==0.5.0
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
uvicorn==0.34.0
